Extracting components from SchedulePage
