+++ src/pages/SchedulePage.js
@@ -1,14 +1,14 @@
-import React, { useState, useEffect, Fragment } from "react";
+import React, { useState, useEffect } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { parseDate, addMinutes, generateScheduleTime } from "../utils";
-import { times, dayToStr } from "../constants";
-import { CustomCheckbox } from "../components/CustomCheckbox";
+import { dayToStr } from "../constants";
+import WishlistCourse from "../components/WishlistCourse";
+import Calendar from "../components/Calendar";
+import ScheduleCourse from "../components/ScheduleCourse";
 
 import styled from "styled-components";
 import SnackBar from "@material-ui/core/Snackbar";
 import { grey } from "@material-ui/core/colors";
-import { FormControlLabel } from "@material-ui/core";
 
 import * as wishlistActions from "../redux/modules/wishlist";
 import * as courseActions from "../redux/modules/courseSelect";
@@ -17,7 +17,6 @@ function SchedulePage({
   year,
   semester,
   wishlist,
-  clearWishlist, // eslint-disable-line no-unused-vars
   wishlistCourse,
   scheduled,
   toggleCourseSelect,
@@ -89,6 +88,15 @@ function SchedulePage({
     })();
   }, [year, semester, scheduled]);
 
+  const handleClearSchedule = () => {
+    clearSchedule({ year, semester });
+    setCheckboxes({});
+    localStorage.setItem(
+      `${year}-${semester}-checkbox-state`,
+      JSON.stringify({})
+    );
+  };
+
   const handleOnClose = () => {
     setToast({
       open: false,
@@ -148,82 +156,15 @@ function SchedulePage({
     );
   };
 
-  const computeMargin = (startTime) => {
-    const parsedDate = parseDate(startTime);
-    return (
-      (parsedDate.getHours() - 8) * 4 + (parsedDate.getMinutes() / 60) * 4 + 1
-    );
-  };
-
-  const _renderCourses = (dayNum, schedule) =>
+  const _renderCourses = (dayNum) =>
     schedule[dayToStr[dayNum]] !== undefined &&
-    Object.values(schedule[dayToStr[dayNum]]).map((course, i) => (
-      <CourseBlock
-        key={i}
-        style={{
-          minHeight: `${(course.minutesDuration / 60) * 4}rem`,
-          marginTop: `${computeMargin(course.beginDate)}rem`,
-        }}
-      >
-        <TextContainer>
-          <div className="courseCode">
-            {`${course.subject.code}-${course.subject.school} ${course.deptCourseId}-${course.sectionCode}`}
-          </div>
-        </TextContainer>
-        <div className="time">
-          {`${parseDate(course.beginDate).toLocaleTimeString([], {
-            hour: "2-digit",
-            minute: "2-digit",
-          })}
-              - ${addMinutes(
-                parseDate(course.beginDate),
-                course.minutesDuration
-              ).toLocaleTimeString([], {
-                hour: "2-digit",
-                minute: "2-digit",
-              })}`}
-        </div>
-        <div className="location">{course.location}</div>
-      </CourseBlock>
-    ));
-
-  const days = [
-    "Sunday",
-    "Monday",
-    "Tuesday",
-    "Wednesday",
-    "Thursday",
-    "Friday",
-    "Saturday",
-  ];
+    Object.values(schedule[dayToStr[dayNum]]).map((course, i) => {
+      return <ScheduleCourse course={course} key={i} />;
+    });
 
   return (
     <Container>
-      <Calendar>
-        <TimeGrid>
-          {times.map((time, i) => {
-            return <Time key={i}> {time} </Time>;
-          })}
-        </TimeGrid>
-        <CourseCalendar>
-          {days.map((day, i) => {
-            //ignoring Saturday and Sunday for now
-            return day === "Sunday" || day === "Saturday" ? (
-              <Fragment key={day} />
-            ) : (
-              <CalendarDay key={day}>
-                {day}
-                {_renderCourses(i, schedule)}
-              </CalendarDay>
-            );
-          })}
-          {Array(65)
-            .fill(1)
-            .map((item, i) => {
-              return <CalendarDay key={i} />;
-            })}
-        </CourseCalendar>
-      </Calendar>
+      <Calendar renderCourses={_renderCourses} />
       <div
         style={{
           marginTop: "2rem",
@@ -247,100 +188,28 @@ function SchedulePage({
           ) : (
             wishlist.map((course, i) => {
               return (
-                <WishlistCourse key={i}>
-                  <WishlistTextBox>
-                    <div>{course.name}</div>
-                    <div>
-                      Section: <span>{course.code}</span>
-                    </div>
-                    <div>
-                      Registration No: <span>{course.registrationNumber}</span>
-                    </div>
-                    <div
-                      style={{
-                        marginTop: "1rem",
-                      }}
-                    >
-                      <div>
-                        Type: <span>{course.type}</span>
-                      </div>
-                      <div>
-                        Instructors:{" "}
-                        <span>{course.instructors.join(", ")}</span>
-                      </div>
-                      <div>
-                        Meetings:{" "}
-                        <span>{generateScheduleTime(course.meetings)}</span>
-                      </div>
-                    </div>
-                    <WishlistUtilBox>
-                      <FormControlLabel
-                        value="add"
-                        control={
-                          <CustomCheckbox
-                            checked={
-                              checkboxes[course.registrationNumber] ===
-                              undefined
-                                ? false
-                                : checkboxes[course.registrationNumber]
-                            }
-                            onChange={(e) =>
-                              handleOnChange(
-                                e,
-                                course,
-                                course.registrationNumber
-                              )
-                            }
-                          >
-                            {" "}
-                          </CustomCheckbox>
-                        }
-                        label="Schedule"
-                        labelPlacement="start"
-                        style={{
-                          margin: "0",
-                          color: "black",
-                          backgroundColor: "var(--grey400)",
-                          borderRadius: "5px",
-                          padding: "0 8px",
-                          fontWeight: "bold",
-                        }}
-                      />
-                      <div
-                        role="button"
-                        className="removeButton"
-                        onClick={() => removeCourse(course)}
-                        onKeyDown={() => removeCourse(course)}
-                        tabIndex={0}
-                      >
-                        Remove
-                      </div>
-                    </WishlistUtilBox>
-                  </WishlistTextBox>
-                </WishlistCourse>
+                <WishlistCourse
+                  key={i}
+                  year={year}
+                  semester={semester}
+                  course={course}
+                  checkboxes={checkboxes}
+                  removeCourse={removeCourse}
+                  handleOnChange={handleOnChange}
+                />
               );
             })
           )}
         </WishlistCoursesList>
-        <div>
-          <div
-            style={{ cursor: "pointer", marginTop: "1rem", color: "#bd2f2f" }}
-            onClick={() => {
-              clearSchedule({ year, semester });
-              setCheckboxes({});
-              localStorage.setItem(
-                `${year}-${semester}-checkbox-state`,
-                JSON.stringify({})
-              );
-            }}
-            onKeyPress={() => clearSchedule({ year, semester })}
-            role="button"
-            tabIndex={0}
-          >
-            Clear Schedule
-          </div>
-          {}
-        </div>
+        <ClearScheduleButton
+          style={{ cursor: "pointer", marginTop: "1rem", color: "#bd2f2f" }}
+          onClick={handleClearSchedule}
+          onKeyPress={() => clearSchedule({ year, semester })}
+          role="button"
+          tabIndex={0}
+        >
+          Clear Schedule
+        </ClearScheduleButton>
       </div>
       <SnackBar
         anchorOrigin={{ vertical, horizontal }}
@@ -358,7 +227,6 @@ SchedulePage.propTypes = {
   year: PropTypes.number.isRequired,
   semester: PropTypes.string.isRequired,
   wishlist: PropTypes.arrayOf(PropTypes.object).isRequired,
-  clearWishlist: PropTypes.func.isRequired,
   wishlistCourse: PropTypes.func.isRequired,
   scheduled: PropTypes.arrayOf(PropTypes.object).isRequired,
   toggleCourseSelect: PropTypes.func.isRequired,
@@ -393,113 +261,10 @@ const WishlistCoursesList = styled.div`
   overflow: scroll;
 `;
 
-const WishlistCourse = styled.div`
-  min-height: 15rem;
-  background-color: var(--grey300);
-  border-bottom: 1px solid var(--grey200);
-  border-top: 1px solid var(--grey200);
-`;
-
-const WishlistTextBox = styled.div`
-  padding: 1rem;
-`;
-
-const WishlistUtilBox = styled.div`
-  display: flex;
-  height: 4rem;
-  margin-top: 1.5rem;
-  align-items: center;
-
-  & > .removeButton {
-    cursor: pointer;
-    color: #bd2f2f;
-    font-size: 0.9rem;
-    margin-left: 1rem;
-  }
-`;
-
-const Calendar = styled.div`
-  min-height: 100vh;
-  padding: 1rem;
-  display: flex;
-`;
-
-const CourseCalendar = styled.div`
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(5, 12rem);
-  grid-template-rows: 48px repeat(13, 4rem);
-  @media (max-width: 1000px) {
-    grid-template-columns: 1fr;
-  }
-`;
-
-const CalendarDay = styled.div`
-  width: 100%;
-  min-height: 2vh;
-  padding: 15px;
-  border-bottom: 1px dashed var(--grey400);
-  text-align: center;
-  align-items: center;
-
-  @media (max-width: 1000px) {
-    min-height: 150px;
-  }
-`;
-
-const TimeGrid = styled.div`
-  width: 5%;
-  display: grid;
-  grid-template-columns: auto;
-  grid-template-rows: 40px repeat(13, 4rem);
-`;
-
-const Time = styled.div`
-  text-align: right;
-  font-size: 1rem;
-  padding: 0 5px 0 0;
-`;
-
-const CourseBlock = styled.div`
-  background-color: var(--grey400);
-  width: 10rem;
-  min-height: 4rem;
-  padding: 5px 5px 5px 10px;
-  border-radius: 10px;
-  position: absolute;
-
-  & > .time {
-    font-size: 0.8rem;
-    text-align: center;
-    padding: 0 2px;
-    width: 100%;
-  }
-
-  & > .location {
-    font-size: 0.8rem;
-    text-align: center;
-    padding: 0 2px;
-    width: 100%;
-  }
-`;
-
-const TextContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-
-  & > .courseCode {
-    font-size: 0.9rem;
-    text-align: center;
-    padding: 0 2px;
-    width: 100%;
-  }
-
-  & > .closeButton {
-    cursor: pointer;
-    &:hover {
-      color: var(--grey100);
-    }
-  }
+const ClearScheduleButton = styled.div`
+  cursor: pointer;
+  margin-top: 1rem;
+  color: #bd2f2f;
 `;
 
 const mapStateToProps = (state, props) => ({
